Refactor: Move WorkflowError to models to decouple imports (#445)
